move anvil cost to its own class
